+++ src/Data/Apiary/SProxy.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE KindSignatures #-}
+{-# LANGUAGE DataKinds #-}
+
+module Data.Apiary.SProxy(SProxy(..)) where
+
+import GHC.TypeLits.Compat(Symbol)
+
+data SProxy (k :: Symbol) = SProxy